Fixed for contro;l hub

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/opmode/TeamCode/Assets/Constants/Slide/SlideHeight.java
@@ -4,11 +4,11 @@ import com.acmerobotics.dashboard.config.Config;
 
 @Config
 public class SlideHeight {
-    public static double HighPole = 100.0;
-    public static double MidPole = 75.0;
-    public static double LowPole = 50.0;
-    public static double Ground = 0.0;
+    public static int HighPole = 100;
+    public static int MidPole = 75;
+    public static int LowPole = 50;
+    public static int Ground = 0;
 
-    public static double Max = 100.0;
-    public static double Min = 0.0;
+    public static int Max = 100;
+    public static int Min = 0;
 }
